Moved host info data to GraphicJson
